refactor(core): merge Tree* classes into Tree
Previously it seemed to make sense to have separate classes TreeExpr,
TreeNode and TreeAtom. The main reason for the distinction was so that
TreeAtom could be generically typed with a rtype parameter indicating
the type of the values stored in the atom. This turns out not to be very
useful though and the three separate classes are an unnecessary
complexity.

This commit merges the three classes TreeExpr, TreeNode and TreeAtom
into a single Tree class. The instances of this class have two
attributes value and children. The two possibilities are that value=None
and children is a tuple ot Tree or that value is not None and children
is the empty tuple. This more naturally corresponds to the Rust
implementation in which Tree is represented as an enum with two
variants.