test(core): tweak tests for rust_protosym compat
A few changes are needed in tests and doctests to make them compatible
with rust_protosym. In particular:

- Some doctests print the type like type(expr) and the rust_protosym
  types display differently.
- Some tests assert identity like `Integer(1) is Integer(1)` but
  rust_protosym does not ensure that kind of identity.

Ideally rust_protosym would also ensure identity of objects but it is
difficult to do in pyo3 because of the way that __new__ needs to be
defined. The internal values held in the Rust struct are interned so
more or less the same performance effect is achieved but just the Python
objects are duplicated. Note that Sym which wraps Tree in pure Python
*does* ensure uniqueness so again atoms wrapped in a Sym will be unique.