feat(core): Add SubsFunc
The SubsFunc class provides a what o efficiently compute a substitution
into an expression where there is a need to repeatedly make
substitutions for the same subexpressions of a particular expression.
This is needed for the case when implementing pattern-matching
replacement rules where the rule might need to be applied many times.